move stuff to more packages for tidiness
