+++ distribution/lib/Standard/Table/0.0.0-dev/src/Internal/Add_Row_Number.enso
@@ -15,8 +15,8 @@ from project.Errors import Duplicate_Output_Column_Names
 from project.Internal.Java_Exports import make_long_builder
 
 polyglot java import org.enso.base.arrays.LongArray
-polyglot java import org.enso.table.data.column.storage.LongStorage
-polyglot java import org.enso.table.data.column.storage.LongRangeStorage
+polyglot java import org.enso.table.data.column.storage.numeric.LongStorage
+polyglot java import org.enso.table.data.column.storage.numeric.LongRangeStorage
 
 ## PRIVATE
 add_row_number : Table -> Text -> Integer -> Integer -> (Column_Selector | Vector Text) -> Vector (Text | Sort_Column) | Text | Sort_Column -> Problem_Behavior -> Table
