+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/object/DateBuilder.java
@@ -1,6 +1,6 @@
 package org.enso.table.data.column.builder.object;
 
-import org.enso.table.data.column.storage.DateStorage;
+import org.enso.table.data.column.storage.datetime.DateStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.type.DateType;
 import org.enso.table.data.column.storage.type.StorageType;
