+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/object/DateTimeBuilder.java
@@ -1,6 +1,6 @@
 package org.enso.table.data.column.builder.object;
 
-import org.enso.table.data.column.storage.DateTimeStorage;
+import org.enso.table.data.column.storage.datetime.DateTimeStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.type.DateTimeType;
 import org.enso.table.data.column.storage.type.StorageType;
