+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/cast/ToDateStorageConverter.java
@@ -1,8 +1,8 @@
 package org.enso.table.data.column.operation.cast;
 
 import org.enso.table.data.column.builder.object.DateBuilder;
-import org.enso.table.data.column.storage.DateStorage;
-import org.enso.table.data.column.storage.DateTimeStorage;
+import org.enso.table.data.column.storage.datetime.DateStorage;
+import org.enso.table.data.column.storage.datetime.DateTimeStorage;
 import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.column.storage.type.AnyObjectType;
