+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/cast/ToTimeOfDayStorageConverter.java
@@ -1,9 +1,9 @@
 package org.enso.table.data.column.operation.cast;
 
 import org.enso.table.data.column.builder.object.TimeOfDayBuilder;
-import org.enso.table.data.column.storage.DateTimeStorage;
+import org.enso.table.data.column.storage.datetime.DateTimeStorage;
 import org.enso.table.data.column.storage.Storage;
-import org.enso.table.data.column.storage.TimeOfDayStorage;
+import org.enso.table.data.column.storage.datetime.TimeOfDayStorage;
 import org.enso.table.data.column.storage.type.AnyObjectType;
 
 import java.time.LocalTime;
