+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/UnaryIntegerOp.java
@@ -1,11 +1,9 @@
 package org.enso.table.data.column.operation.map;
 
-import org.enso.table.data.column.storage.LongStorage;
+import org.enso.table.data.column.storage.numeric.LongStorage;
 import org.enso.table.data.column.storage.Storage;
-import org.enso.table.util.BitSets;
 
 import java.util.BitSet;
-import java.util.function.Function;
 
 /** An operation that takes a single argument of some type and returns an integer. */
 public abstract class UnaryIntegerOp<T, I extends Storage<T>> extends UnaryMapOperation<T, I> {
