+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/map/numeric/LongIsInOp.java
@@ -2,8 +2,7 @@ package org.enso.table.data.column.operation.map.numeric;
 
 import org.enso.base.polyglot.NumericConverter;
 import org.enso.table.data.column.operation.map.SpecializedIsInOp;
-import org.enso.table.data.column.storage.AbstractLongStorage;
-import org.enso.table.data.column.storage.LongStorage;
+import org.enso.table.data.column.storage.numeric.AbstractLongStorage;
 
 import java.util.HashSet;
 import java.util.List;
