+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/datetime/TimeOfDayStorage.java
@@ -1,9 +1,11 @@
-package org.enso.table.data.column.storage;
+package org.enso.table.data.column.storage.datetime;
 
 import org.enso.table.data.column.builder.object.Builder;
 import org.enso.table.data.column.builder.object.TimeOfDayBuilder;
 import org.enso.table.data.column.operation.map.MapOpStorage;
 import org.enso.table.data.column.operation.map.datetime.DateTimeIsInOp;
+import org.enso.table.data.column.storage.ObjectStorage;
+import org.enso.table.data.column.storage.SpecializedStorage;
 import org.enso.table.data.column.storage.type.StorageType;
 import org.enso.table.data.column.storage.type.TimeOfDayType;
